first pkt error stil here

+++ srcs/headers.c
@@ -33,7 +33,7 @@ void                    genhdr(struct ip *ip, struct icmp *icmp)
 
 	g_data.stat.tsin = gettimestamp_ms(0);
 
-    ip->ip_v = 4;
+	ip->ip_v = 4;
 	ip->ip_hl = sizeof(*(ip)) >> 2;
 	ip->ip_tos = 0;
 	ip->ip_len = PACKET_SIZE;
@@ -45,10 +45,10 @@ void                    genhdr(struct ip *ip, struct icmp *icmp)
 	ip->ip_src.s_addr = INADDR_ANY;
 	ip->ip_dst.s_addr = ((const struct sockaddr_in*)g_data.info->ai_addr)->sin_addr.s_addr;
 
-    icmp->icmp_type = ICMP_ECHO;
-    icmp->icmp_code = 0;
-    icmp->icmp_id = (unsigned short)g_data.pid;
-    icmp->icmp_cksum = 0;
-    icmp->icmp_seq = (unsigned short)g_data.stat.seq++;
-    icmp->icmp_cksum = calculatechecksum((unsigned short*)icmp);
+	icmp->icmp_type = ICMP_ECHO;
+	icmp->icmp_code = 0;
+	icmp->icmp_id = g_data.pid;
+	icmp->icmp_cksum = 0;
+	icmp->icmp_seq = (unsigned short)g_data.stat.seq++;
+	icmp->icmp_cksum = calculatechecksum((unsigned short*)icmp);
 }
